Refactor controllers
Por fin aprendí a usar closures en php :)

+++ app/Aluc/Service/AdministradorSrv.php
@@ -1,6 +1,7 @@
 <?php
 namespace Aluc\Service;
 
+use Aluc\Models\Lector;
 use Aluc\Models\Moderador;
 use Aluc\Views\AdministradorView;
 use Aluc\Views\LectorQrView;
@@ -25,36 +26,17 @@ class AdministradorSrv {
         static::$view_administrador = AdministradorView::getInstance();
     }
 
-    private static function admin_do($func, $view_success, $view_error) {
-        $view = null;
+    private static function admin_do($func, $error) {
         try {
             if (Tools::check_session('admin')) {
-                if (Tools::check_method('post')) {
-                    $func();
-                }
-                $view = $view_success();
+                $func();
             } else {
-                $view = self::$view_general->error404();
+                self::$view_general
+                    ->error404()
+                    ->render();
             }
         } catch (\Exception $e) {
-            $view = $view_error();
-        } finally {
-            $view->render();
-        }
-    }
-
-    private static function admin_go($view_go, $view_error) {
-        $view = null;
-        try {
-            if (Tools::check_session('admin')) {
-                $view = $view_go();
-            } else {
-                $view = self::$view_general->error404();
-            }
-        } catch (\Exception $e) {
-            $view  = $view_error();
-        } finally {
-            $view->render();
+            $error($e);
         }
     }
 
@@ -64,14 +46,19 @@ class AdministradorSrv {
      * Muestra la página principal del administrador
      * desde la cual tiene acceso a varias herramientas
      * de administración.
+     * @param $data
      */
-    public static function home() {
-        self::admin_go(
-            function () {
-                return self::$view_administrador->home();
+    public static function home($data) {
+        self::admin_do(
+            function () use ($data){
+                self::$view_administrador
+                    ->home()
+                    ->render();
             },
-            function () {
-                return self::$view_general->error404();
+            function ($e) {
+                self::$view_general
+                    ->error404()
+                    ->render();
             }
         );
     }
@@ -81,14 +68,19 @@ class AdministradorSrv {
      *
      * Muestra una lista de todos los moderadores.
      * Sólo el administrador tiene acceso.
+     * @param $data
      */
-    public static function moderadores() {
-        self::admin_go(
-            function () {
-                return self::$view_moderador->listAll();
+    public static function moderadores($data) {
+        self::admin_do(
+            function () use ($data) {
+                self::$view_moderador
+                    ->listAll()
+                    ->render();
             },
-            function () {
-                return self::$view_general->error404();
+            function ($e) {
+                self::$view_general
+                    ->error404()
+                    ->render();
             }
         );
     }
@@ -100,20 +92,25 @@ class AdministradorSrv {
      * Crea un nuevo moderador.
      * La petición se debe hacer vía post.
      * Sólo un administrador puede realizar esta acción.
+     * @param $data
      */
-    public static function moderadores_nuevo() {
-        self::admin_do(
-            function () {
-                $id = Tools::clean_string($_POST['id']);
-                $laboratorio_id = Tools::clean_string($_POST['laboratorio_id']);
-                $moderador = Moderador::getNewInstace($id, $laboratorio_id);
-                $moderador->save();
+    public static function moderadores_nuevo($data) {
+        self::admin_do (
+            function () use ($data) {
+                if (!empty($data) and Tools::check_method('post')) {
+                    $id = $data['id'];
+                    $laboratorio_id = $data['laboratorio_id'];
+                    $moderador = Moderador::getNewInstace($id, $laboratorio_id);
+                    $moderador->save();
+                }
+                self::$view_moderador
+                    ->listAll()
+                    ->render();
             },
-            function () {
-                return self::$view_moderador->listAll();
-            },
-            function () {
-                return self::$view_general->error404();
+            function ($e) {
+                self::$view_general
+                    ->error404()
+                    ->render();
             }
         );
     }
@@ -124,21 +121,26 @@ class AdministradorSrv {
      * Editar la información de un moderador.
      * La petición se debe hacer vía post.
      * Sólo un administrador puede realizar esta acción.
+     * @param $data
      */
-    public static function moderadores_actualizar() {
+    public static function moderadores_actualizar($data) {
         self::admin_do(
-            function () {
-                $id = Tools::clean_string($_POST['id']);
-                $laboratorio_id = Tools::clean_string($_POST['laboratorio_id']);
-                $moderador = Moderador::getInstance($id);
-                $moderador->id_laboratorio = $laboratorio_id;
-                $moderador->save();
+            function () use ($data) {
+                if (!empty($data) and Tools::check_method('post')) {
+                    $id = $data['id'];
+                    $laboratorio_id = $data['laboratorio_id'];
+                    $moderador = Moderador::getInstance($id);
+                    $moderador->id_laboratorio = $laboratorio_id;
+                    $moderador->save();
+                }
+                self::$view_moderador
+                    ->listAll()
+                    ->render();
             },
-            function () {
-                return self::$view_moderador->listAll();
-            },
-            function () {
-                return self::$view_general->error404();
+            function ($e) {
+                self::$view_general
+                    ->error404()
+                    ->render();
             }
         );
     }
@@ -150,19 +152,24 @@ class AdministradorSrv {
      * Elimina un moderador del sistema.
      * La petición se debe hacer vía post.
      * Sólo un administrador puede realizar esta acción.
+     * @param $data
      */
-    public static function moderadores_eliminar() {
+    public static function moderadores_eliminar($data) {
         self::admin_do(
-            function () {
-                $id = Tools::clean_string($_POST['id']);
-                $moderador = Moderador::getInstance($id);
-                $moderador->delete();
+            function () use ($data) {
+                if (!empty($data) and Tools::check_method('post')) {
+                    $id = $data['id'];
+                    $moderador = Moderador::getInstance($id);
+                    $moderador->delete();
+                }
+                self::$view_moderador
+                    ->listAll()
+                    ->render();
             },
-            function () {
-                return self::$view_moderador->listAll();
-            },
-            function () {
-                return self::$view_general->error404();
+            function ($e) {
+                self::$view_general
+                    ->error404()
+                    ->render();
             }
         );
     }
@@ -172,14 +179,19 @@ class AdministradorSrv {
      *
      * Muestra una lista de todos los lectores QR.
      * Sólo el administrador tiene acceso.
+     * @param $data
      */
-    public static function lectores() {
-        self::admin_go(
-            function () {
-                return self::$view_lector_qr->listAll();
+    public static function lectores($data) {
+        self::admin_do(
+            function () use ($data) {
+                self::$view_lector_qr
+                    ->listAll()
+                    ->render();
             },
             function () {
-                return self::$view_general->error404();
+                self::$view_general
+                    ->error404()
+                    ->render();
             }
         );
     }
@@ -191,21 +203,26 @@ class AdministradorSrv {
      * Crea un nuevo lector QR.
      * La petición se debe hacer vía post.
      * Sólo un administrador puede realizar esta acción.
+     * @param $data
      */
-    public static function lectores_nuevo() {
+    public static function lectores_nuevo($data) {
         self::admin_do(
-            function () {
-                $ip = Tools::clean_string($_POST['ip']);
-                $mac = Tools::clean_string($_POST['mac']);
-                $laboratorio_id = Tools::clean_string($_POST['laboratorio_id']);
-                $lector = Lector::getNewInstance($ip, $mac, $laboratorio_id);
-                $lector->save();
+            function () use ($data) {
+                if (!empty($data) and Tools::check_method('post')) {
+                    $ip = $data['ip'];
+                    $mac = $data['mac'];
+                    $laboratorio_id = $data['laboratorio_id'];
+                    $lector = Lector::getNewInstance($ip, $mac, $laboratorio_id);
+                    $lector->save();
+                }
+                self::$view_lector_qr
+                    ->listAll()
+                    ->render();
             },
-            function () {
-                return self::$view_lector_qr->listAll();
-            },
-            function () {
-                return self::$view_general->error404();
+            function ($e) {
+                self::$view_general
+                    ->error404()
+                    ->render();
             }
         );
     }
@@ -218,29 +235,34 @@ class AdministradorSrv {
      * token para el lector.
      * La petición se debe hacer vía post.
      * Sólo un administrador puede realizar esta acción.
+     * @param $data
      */
-    public static function lectores_actualizar() {
+    public static function lectores_actualizar($data) {
         self::admin_do(
-            function () {
-                $id = Tools::clean_string($_POST['id']);
-                $ip = Tools::clean_string($_POST['ip']);
-                $mac = Tools::clean_string($_POST['mac']);
-                $laboratorio_id = Tools::clean_string($_POST['laboratorio_id']);
-                $new_token = strtolower(Tools::clean_string($_POST['new_token'])) === 'true';
-                $lector = Lector::getInstance($id);
-                $lector->ip = $ip;
-                $lector->mac = $mac;
-                $lector->laboratorio_id = $laboratorio_id;
-                if ($new_token) {
-                    $lector->renovarToken();
+            function () use ($data) {
+                if (!empty($data) and Tools::check_method('post')) {
+                    $id = $data['id'];
+                    $ip = $data['ip'];
+                    $mac = $data['mac'];
+                    $laboratorio_id = $data['laboratorio_id'];
+                    $new_token = strtolower($data['new_token']) === 'true';
+                    $lector = Lector::getInstance($id);
+                    $lector->ip = $ip;
+                    $lector->mac = $mac;
+                    $lector->laboratorio_id = $laboratorio_id;
+                    if ($new_token) {
+                        $lector->renovarToken();
+                    }
+                    $lector->save();
                 }
-                $lector->save();
+                self::$view_lector_qr
+                    ->listAll()
+                    ->render();
             },
-            function () {
-                return self::$view_lector_qr->listAll();
-            },
-            function () {
-                return self::$view_general->error404();
+            function ($e) {
+                self::$view_general
+                    ->error404()
+                    ->render();
             }
         );
     }
@@ -249,20 +271,25 @@ class AdministradorSrv {
      * Elimina un lector del sistema.
      * La petición se debe hacer vía post.
      * Sólo un administrador puede realizar esta acción.
+     * @param $data
      */
-    public static function lectores_eliminar() {
+    public static function lectores_eliminar($data) {
         self::admin_do(
-            function () {
-                $id = Tools::clean_string($_POST['id']);
-                $laboratorio_id = Tools::clean_string($_POST['laboratorio_id']);
-                $lector = Lector::getInstance($id, $laboratorio_id);
-                $lector->delete();
+            function () use ($data) {
+                if (!empty($data) and Tools::check_method('post')) {
+                    $id = $data['id'];
+                    $laboratorio_id = $data['laboratorio_id'];
+                    $lector = Lector::getInstance($id, $laboratorio_id);
+                    $lector->delete();
+                }
+                self::$view_lector_qr
+                    ->listAll()
+                    ->render();
             },
             function () {
-                return self::$view_lector_qr->listAll();
-            },
-            function () {
-                return self::$view_general->error404();
+                self::$view_general
+                    ->error404()
+                    ->render();
             }
         );
     }
